+++ src/features/game1/container/index.js
@@ -1,4 +1,5 @@
 import FaceRecognitionContainer from "./FaceRecognitionContainer";
 import Game1Container from "./Game1Container";
+import MapContainer from "./MapContainer";
 
-export { FaceRecognitionContainer, Game1Container };
+export { FaceRecognitionContainer, Game1Container, MapContainer };
