+++ src/features/game1/presenter/index.js
@@ -1,6 +1,7 @@
+import DefaultMap from "./DefaultMap";
+import FaceDirectionRecord from "./FaceDirectionRecord";
 import FaceRecognition from "../presenter/FaceRecognition";
 import Game1 from "../presenter/Game1";
 import Map from "../presenter/Map";
-import FaceDirectionRecord from "./FaceDirectionRecord";
 
-export { Game1, Map, FaceRecognition, FaceDirectionRecord };
+export { DefaultMap, FaceDirectionRecord, FaceRecognition, Game1, Map };
