+++ src/features/main/presenter/index.js
@@ -1,0 +1,5 @@
+import GameGuide from "./GameGuide";
+import Main from "./Main";
+import Page from "./Page";
+
+export { GameGuide, Main, Page };
